LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/pseudo-elements-013-expected.txt


FAIL Initial text-decoration-thickness for highlight pseudo assert_equals: expected "0px" but got ""
FAIL text-decoration-thickness for highlight pseudo depending on container assert_equals: expected "30px" but got ""
FAIL text-decoration-thickness for highlight pseudo depending on container only defined in a query assert_equals: expected "30px" but got ""
